Simplified the code to remove use of channels and just use locks
and a waitGroup

-- OLD code benchmark --
```
anthony:~/go/src/github.com/grafana/metrictank/idx/memory$ go test -v -race -run none -bench TimeLimiter -benchmem
goos: linux
goarch: amd64
pkg: github.com/grafana/metrictank/idx/memory
BenchmarkTimeLimiterWaitNotLimited-4      500000              3245 ns/op              96 B/op          1 allocs/op
BenchmarkTimeLimiterAddNotLimited-4      1000000              1589 ns/op               0 B/op          0 allocs/op
BenchmarkTimeLimiterAddLimited-4         1000000              1552 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/grafana/metrictank/idx/memory        5.857s
```
-- NEW code benchmark --
```
anthony:~/go/src/github.com/grafana/metrictank/idx/memory$ go test -v -race -run none -bench TimeLimiter -benchmem
goos: linux
goarch: amd64
pkg: github.com/grafana/metrictank/idx/memory
BenchmarkTimeLimiterWaitNotLimited-4    10000000               129 ns/op               0 B/op          0 allocs/op
BenchmarkTimeLimiterAddNotLimited-4      3000000               497 ns/op               0 B/op          0 allocs/op
BenchmarkTimeLimiterAddLimited-4         3000000               530 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/grafana/metrictank/idx/memory        6.574s
b6d0437
```