@Dieterbe 2018-10-10
looks good.
I actually have done something similar but took a different approach: #1088 
the goal there was mainly to make unit testing robust, which was done by removing concurrency features (which we don't need anyway) which in turn also meant the accounting could be made more accurate