@Dieterbe 2018-10-10
shouldn't the benchmarks include a call to Wait? not sure what the use is of only testing adds but not waits, as we would never use TimeLimiter that way.