@Dieterbe 2018-10-11
IMHO, we should add `Wait` calls to the benchmarks, add the benchmarks to #1088 and merge 1088 instead of this PR because it is simpler, more accurate accounting and fixes the broken unit tests.